Can save activity with hibernate / JPA

+++ de.bruenni.sabbat.location.infrastructure/src/main/postgres/init/postgres_init_tables.sql
@@ -9,7 +9,7 @@ BEGIN;
     CREATE TYPE domain_event_type AS ENUM ('activity');
 
     CREATE TABLE IF NOT EXISTS loc.activity (
-        id SERIAL PRIMARY KEY,
+        id BIGSERIAL PRIMARY KEY,
         uuid VARCHAR(40) not null UNIQUE,    -- activity uuid created by clients (Oruxmaps e.g.)
         userid VARCHAR(255) not null,
         title VARCHAR(255) not null,
@@ -19,30 +19,28 @@ BEGIN;
 
     CREATE UNIQUE INDEX IF NOT EXISTS idx_activity_uuid ON loc.activity (uuid);
 
-
-
     ---------------------------------------
 
     CREATE TABLE IF NOT EXISTS loc.activityrelation (
-        id SERIAL PRIMARY KEY,
-        activityid1 SERIAL not null REFERENCES loc.activity (id) ON DELETE CASCADE,
-        activityid2 SERIAL not null REFERENCES loc.activity (id) ON DELETE CASCADE
+        id BIGSERIAL PRIMARY KEY,
+        activityid1 BIGSERIAL not null REFERENCES loc.activity (id) ON DELETE CASCADE,
+        activityid2 BIGSERIAL not null REFERENCES loc.activity (id) ON DELETE CASCADE
     );
 
     --ALTER TABLE IF EXISTS loc.activityrelation
-      --  ADD COLUMN IF NOT EXISTS activityids SERIAL [ column_constraint [ ... ] ]
+      --  ADD COLUMN IF NOT EXISTS activityids BIGSERIAL [ column_constraint [ ... ] ]
 
     -- time series events coming in from CEP with new distances between activities --
     CREATE TABLE IF NOT EXISTS loc.activityrelationevents (
-        id SERIAL PRIMARY KEY,
-        aggregateid SERIAL not null REFERENCES loc.activityrelation (id) ON DELETE CASCADE,
+        id BIGSERIAL PRIMARY KEY,
+        aggregateid BIGSERIAL not null REFERENCES loc.activityrelation (id) ON DELETE CASCADE,
         document JSONB not null
     );
 
         -- -- DOMAIN events --
     CREATE TABLE IF NOT EXISTS loc.domainevents (
-        id SERIAL PRIMARY KEY,
-        aggregateid SERIAL not null,
+        id BIGSERIAL PRIMARY KEY,
+        aggregateid BIGSERIAL not null,
         type domain_event_type not null,
         document JSONB not null
         -- FOREIGN KEY (aggregateid) REFERENCES loc.activity (id) ON DELETE CASCADE,
